+++ examples/hello/libraries/same-things-2/same-thing.nix
@@ -1,0 +1,11 @@
+{ base }:
+base.mkDerivation {
+  name = "same-thing";
+  src = ./.;
+  builder = builtins.toFile "builder.sh" ''
+    source $stdenv/setup
+    ls $src
+    mkdir -p $out
+    cp $src/same.txt $out/same.txt
+  '';
+}