+++ test/filter-source.nix
@@ -1,0 +1,5 @@
+assertMsg: drv1: drv2:
+assert assertMsg
+  (drv1 == drv2)
+  "same-thing-1 and same-thing-2 differes, which they should not!";
+{ }