+++ examples/dependencies/child/clients/nooo/nooo.nix
@@ -1,0 +1,5 @@
+{ base, parents }:
+base.languages.ewokese.mkChild {
+  name = "Luke";
+  parents = builtins.concatStringsSep ", " (builtins.map (p: "${p.name} version: ${p.version or "not-versioned"}") parents);
+}