+++ test.nix
@@ -5,6 +5,7 @@ let
     hello = import ./examples/hello/project.nix;
     deployment = import ./deployment/project.nix;
     documentation = import ./examples/documentation/project.nix;
+    dependencies = import ./examples/dependencies/child/project.nix;
   };
 
   pkgs = tests.hello.nixpkgs;
@@ -19,6 +20,7 @@ let
     combinedDeployment = import ./test/deployment.nix tests.hello.matrix.combinedDeployment.deploy pkgs.lib.assertMsg;
     nestedComponents = import ./test/nested-components.nix pkgs tests.hello;
     sameWhenFiltered = import ./test/filter-source.nix pkgs.lib.assertMsg tests.hello.matrix.sameThing1 tests.hello.matrix.sameThing2;
+    versionTest = import ./test/dependencies.nix tests.dependencies.matrix.nooo pkgs.lib.assertMsg;
   };
 in
 (mappedTests // {