+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Change sphinx and sphinx_rtd_theme to sphinx4 and sphinx4_rtd_theme respectively in
+  order to make sure that we don't invalidate the build for anything that needs Sphinx.
+
 ### Changed
 - "gitingore.nix" for src. This will make all of our own components ignore files that
   are inside gitignore in src.