+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for a project when importing Nedryland. 
 - Shellcheck is now available as part of the `ci` component.
 - nix-lint is now available as part of the `ci` component.
+- version argument to `mkProject`. It will be available on the project.
 
 ## [1.2.0] - 2021-08-31